refactor(booking): split BookingSidebar into modular context-based architecture
