+++ apps/web/src/app/properties/[slug]/components/booking-sidebar/index.tsx
@@ -1,0 +1,2 @@
+export { BookingSidebar } from "./booking-sidebar";
+export type { Room } from "./context/booking-sidebar-context";
