+++ apps/web/src/app/properties/[slug]/components/booking-sidebar/utils/date-formatters.ts
@@ -1,0 +1,40 @@
+/**
+ * Format date for display in date picker button
+ */
+export const formatDateDisplay = (date: Date | undefined): string => {
+  if (!date) return "Select date";
+  return date.toLocaleDateString("en-US", {
+    month: "short",
+    day: "numeric",
+  });
+};
+
+/**
+ * Format date for booking API (YYYY-MM-DD)
+ */
+export const formatDateForBooking = (date: Date): string => {
+  const year = date.getFullYear();
+  const month = String(date.getMonth() + 1).padStart(2, "0");
+  const day = String(date.getDate()).padStart(2, "0");
+  return `${year}-${month}-${day}`;
+};
+
+/**
+ * Format date for conflict display
+ */
+export const formatConflictDate = (date: Date): string => {
+  return date.toLocaleDateString("en-US", {
+    month: "short",
+    day: "numeric",
+    year: "numeric",
+  });
+};
+
+/**
+ * Normalize date to midnight for comparison
+ */
+export const normalizeDateToMidnight = (date: Date): Date => {
+  const normalized = new Date(date);
+  normalized.setHours(0, 0, 0, 0);
+  return normalized;
+};
